+++ ReleaseNotes.md
@@ -1,4 +1,15 @@
-## 2.2.4.23, Pralix, founder of a new House
+## 2.3.0.0, Quendalon
+
+### Features & changes
+
+- [technical] migration of select helper to selectOptions (2 remaining)
+
+### Bug fixes
+
+- Some PC sheet's text editor fields were not setup properly (decrepitude, twilight scars, sigil)
+- Default tab of scriptorium is back to "Reading"
+
+## 2.2.4.24, Pralix, founder of a new House
 
 ### Dependencies updates